ffmpeg 16bit signed integers
sounds and shit ftw c

+++ other/ffmpeg/csrc/include/stuff.h
@@ -1,1 +1,3 @@
-int decode_audio_file(const char* path, const int sample_rate, double** data, int* size); 
+#include <stdint.h>
+
+int decode_audio_file(const char* path, const int sample_rate, int16_t** data, int* size);
